+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## (IN PROGRESS)
 
+## [1.2.1](https://github.com/folio-org/ui-receiving/tree/v1.2.1) (2020-11-10)
+[Full Changelog](https://github.com/folio-org/ui-receiving/compare/v1.2.0...v1.2.1)
+
 * Fix Piece displays as connected to an item when connected item record has been deleted. Refs UIREC-102
 * Orders and receiving - want to see that an order is closed in search result list. Refs UIOR-620
-* Display warning modal when user trying to receive against a closed Order. Refs UIREC-101.
+* Display warning modal when user trying to receive against a closed Order. Refs UIREC-100.
 
 ## [1.2.0](https://github.com/folio-org/ui-receiving/tree/v1.2.0) (2020-10-09)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v1.1.0...v1.2.0)